LinDistFlow.jl

LinDistFlow builds the linear distflow constraints using JuMP. Currently only the single phase model is supported, but the multiphase unbalanced model is under progress.

The intent of this package is to allow users to build mathematical programs that include LinDistFlow constraints. No objective is added to the JuMP model in this package and so solving any problem defined by the constraints built by LinDistFlow.jl is a feasibility problem. Dictionaries of constraints are provided so that one can delete and/or modify the base constraints to fit their problem.

Inputs

There are two methods for creating Inputs:

Using openDSS files

Providing the network topology

LinDistFlow.Inputs — Method

Inputs(
    dssfilepath::String, 
    substation_bus::String;
    Pload::AbstractDict=Dict(), 
    Qload::AbstractDict=Dict(), 
    Sbase=1, 
    Vbase=1, 
    v0, 
    v_lolim=0.95, 
    v_uplim=1.05,
    Ntimesteps=1, 
    P_up_bound=1e4,
    Q_up_bound=1e4,
    P_lo_bound=-1e4,
    Q_lo_bound=-1e4,
)

Inputs constructor that parses a openDSS file for the network. If Pload and Qload are not provided then the loads are also parsed from the openDSS file.

LinDistFlow.Inputs — Method

Inputs(
    edges::Array{Tuple}, 
    linecodes::Array{String}, 
    linelengths::Array{Float64}, 
    phases::Vector{Vector},
    substation_bus::String;
    Pload, 
    Qload, 
    Sbase=1, 
    Vbase=1, 
    Zdict, 
    v0, 
    v_lolim=0.95, 
    v_uplim=1.05,
    Ntimesteps=1, 
    P_up_bound=1e4,
    Q_up_bound=1e4,
    P_lo_bound=-1e4,
    Q_lo_bound=-1e4,
)

Lowest level Inputs constructor (the only one that returns the Inputs struct).

Note

The real and reactive loads provided are normalized using Sbase.

Both of the Inputs functions return a mutable Inputs struct:

LinDistFlow.Inputs — Type

mutable struct Inputs{T<:Phases} <: AbstractInputs
    edges::Array{Tuple, 1}
    linecodes::Array{String, 1}
    linelengths::Array{Float64, 1}
    busses::Array{String}
    phases::Vector{Vector}
    substation_bus::String
    Pload::Dict{String, Any}
    Qload::Dict{String, Any}
    Sbase::Real
    Vbase::Real
    Ibase::Real
    Zdict::Dict{String, Dict{String, Any}}
    v0::Real
    v_lolim::Real
    v_uplim::Real
    Zbase::Real
    Ntimesteps::Int
    pf::Float64
    Nnodes::Int
    P_up_bound::Float64
    Q_up_bound::Float64
    P_lo_bound::Float64
    Q_lo_bound::Float64
    phases_into_bus::Dict{String, Vector{Int}}
end

Building a Model

The build_ldf! function takes a JuMP.Model and Inputs struct as its two arguments and adds the variables and constraints:

LinDistFlow.build_ldf! — Function

build_ldf!(m::JuMP.AbstractModel, p::Inputs)

Add variables and constraints to m using the values in p. Calls the following functions:

add_variables(m, p)
constrain_power_balance(m, p)
constrain_substation_voltage(m, p)
constrain_KVL(m, p)
constrain_loads(m, p)

Variables

Let m be the JuMP.Model provided by the user, then the variables can be accessed via:

m[:vsqrd] voltage magnitude squared, indexed on busses, (phases), time

m[:Pⱼ], m[:Qⱼ] net real, reactive power injection, indexed on busses, (phases), time

m[:Pᵢⱼ], m[:Qᵢⱼ] net real, reactive line flow, indexed on edges, (phases), time

After a model has been solved using JuMP.optimize! variable values can be extracted with JuMP.value. For more see Getting started with JuMP.

Note

Single phase models do not have a phase index

Accessing and Modifying Constraints

Let the JuMP.Model provided by the user be called m. All constraints are stored in m[:cons] as anonymous constraints.

Power Injections

LinDistFlow.jl uses the convention that power injections are positive (and loads are negative). If no load is provided for a given bus (and phase) then the real and reactive power injections at that bus (and phase) are set to zero with an equality constraint.

All power injection constraints are stored in m[:cons][:injection_equalities]. The constraints are indexed in the following order:

by bus name (string), as provided in Inputs.busses;

by :P or :Q for real and reactive power respectively;

by phase number (integer); and

by time (integer).

For example, m[:cons][:injection_equalities]["680"][:P][2][1] contains the constraint reference for the power injection equality constraint for bus "680", real power, on phase 2, in time step 1.

If one wished to replace any constraint one must first delete the constraint using the delete function. For example:

delete(m, m[:cons][:injection_equalities]["680"][:P][1])

Note that the time index was not provided in the delete command in this example, which implies that the equality constraints for all time steps were deleted. One can also delete individual time step constraints by providing the time index.

The deleted constraints can then be replaced with a new set of constraints. For example:

m[:cons][:injection_equalities]["680"][:P][1] = @constraint(m, [t in 1:p.Ntimesteps],
    m[:Pⱼ]["680",1,t] == -1e3 / p.Sbase
)

where p is short for "parameters" and is the Inputs struct for the problem of interest. Note that it is not necessary to store the new constraints in the m[:cons][:injection_equalities].

See the JuMP documentation for more on deleting constraints.