LinDistFlow.jl

LinDistFlow builds the linear distflow constraints using JuMP. The intent of this package is to allow users to build mathematical programs that include LinDistFlow constraints. No objective is added to the JuMP model in this package and so solving any problem defined by the constraints built by LinDistFlow.jl is a feasibility problem. Dictionaries of constraints are provided so that one can delete and/or modify the base constraints to fit their problem.

Inputs

There are two methods for creating Inputs:

Using openDSS files

Providing the network topology

Missing docstring.

Missing docstring for Inputs(::String, ::String). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Inputs(::Array{Tuple}, ::Array{String}, ::Array{Float64}, ::Vector{Vector}, ::String). Check Documenter's build log for details.

Both of the Inputs functions return a mutable Inputs struct:

Missing docstring.

Missing docstring for Inputs. Check Documenter's build log for details.

Building a Model

The build_ldf! function takes a JuMP.Model and Inputs struct as its two arguments and adds the variables and constraints:

LinDistFlow.build_ldf! — Function

build_ldf!(m::JuMP.AbstractModel, p::Inputs)

Add variables and constraints to m using the values in p. Calls the following functions:

add_variables(m, p)
constrain_power_balance(m, p)
constrain_substation_voltage(m, p)
constrain_KVL(m, p)
constrain_loads(m, p)

Variables

Let m be the JuMP.Model provided by the user, then the variables can be accessed via:

m[:vsqrd] voltage magnitude squared, indexed on busses, (phases), time

m[:Pj], m[:Qj] net real, reactive power injection, indexed on busses, (phases), time

m[:Pij], m[:Qij] net real, reactive line flow, indexed on edges, (phases), time

After a model has been solved using JuMP.optimize! variable values can be extracted with JuMP.value. For more see Getting started with JuMP.

Note

Single phase models do not have a phase index

Accessing and Modifying Constraints

Let the JuMP.Model provided by the user be called m. All constraints are stored in m[:cons] as anonymous constraints.

Power Injections

LinDistFlow.jl uses the convention that power injections are positive (and loads are negative). If no load is provided for a given bus (and phase) then the real and reactive power injections at that bus (and phase) are set to zero with an equality constraint.

All power injection constraints are stored in m[:cons][:injection_equalities]. The constraints are indexed in the following order:

by bus name (string), as provided in Inputs.busses;

by :P or :Q for real and reactive power respectively;

by phase number (integer); and

by time (integer).

For example, m[:cons][:injection_equalities]["680"][:P][2][1] contains the constraint reference for the power injection equality constraint for bus "680", real power, on phase 2, in time step 1.

If one wished to replace any constraint one must first delete the constraint using the delete function. For example:

delete(m, m[:cons][:injection_equalities]["680"][:P][1])

Note that the time index was not provided in the delete command in this example, which implies that the equality constraints for all time steps were deleted. One can also delete individual time step constraints by providing the time index.

The deleted constraints can then be replaced with a new set of constraints. For example:

m[:cons][:injection_equalities]["680"][:P][1] = @constraint(m, [t in 1:p.Ntimesteps],
    m[:Pj]["680",1,t] == -1e3 / p.Sbase
)

where p is short for "parameters" and is the Inputs struct for the problem of interest. Note that it is not necessary to store the new constraints in the m[:cons][:injection_equalities].

See the JuMP documentation for more on deleting constraints.